decompiler: Fix joining variables in different spaces
I needed to join a register and a variable in memory space, and ran into
problems because this code wasn't expecting the different varnodes to be
in different spaces with different delays.

I think that delaying building the PIECE and SUBPIECE operations is an
optimization, so I tried to keep that optimization in place.